Improve move speed
Cache current tree properties to prevent repeated calculation

+++ bart_playground/params.py
@@ -126,7 +126,7 @@ class Tree:
         self.dataX = dataX
         self.vars = vars
         self.thresholds = thresholds
-        self.leaf_vals : NDArray[np.float64] = leaf_vals
+        self.leaf_vals : NDArray[np.float32] = leaf_vals
 
         self.n = n
         self.node_indicators = node_indicators
@@ -215,7 +215,7 @@ class Tree:
                 split_node_counter += 1
         return node_ids
 
-    def evaluate(self, X: Optional[np.ndarray]=None) -> NDArray[np.float64]:
+    def evaluate(self, X: Optional[np.ndarray]=None) -> NDArray[np.float32]:
         """
         Evaluate the tree for a given input data matrix.
 
@@ -667,7 +667,7 @@ class Parameters:
         
         return new_state
 
-    def evaluate(self, X: Optional[np.ndarray]=None, tree_ids:Optional[list[int]]=None, all_except:Optional[list[int]]=None) -> NDArray[np.float64]:
+    def evaluate(self, X: Optional[np.ndarray]=None, tree_ids:Optional[list[int]]=None, all_except:Optional[list[int]]=None) -> NDArray[np.float32]:
         """
         Evaluate the model on the given data.
 
@@ -719,7 +719,7 @@ class Parameters:
             return self.trees[tree_ids[0]].leaf_basis
         return np.hstack([self.trees[tree_id].leaf_basis for tree_id in tree_ids])
 
-    def update_leaf_vals(self, tree_ids : list[int], leaf_vals : NDArray[np.float64]):
+    def update_leaf_vals(self, tree_ids : list[int], leaf_vals : NDArray[np.float32]):
         """
         Update the leaf values of specified trees.
